Modularize quality checks
